// repos/sbs-draft-web/utils/types/types.ts
export enum ViewState {
    DRAFT,
    QUEUE,
    BOARD,
    ROSTER,
    CHAT,
    LEADERBOARD,
}
// ...
export type PlayerStateInfo = {
    playerId: string
    displayName: string
    team: string
    position: string
    ownerAddress: string
    pickNum: number
    round: number
}
// ...
export type leagueProps = {
    leagueId: string | null
    leagueName: string | null
    currentRound: number | null
    currentPickNumber: number | null
    currentDrafter: string | null
    queuedPlayers: PlayerStateInfo[]
    endOfTurnTimestamp: number | null
    startOfTurnTimestamp: number | null
    draftStartTime: number | null
    autopick: boolean
    mostRecentPlayerDrafted: mostRecentPlayerProps | null
    leagueStatus: null | string
    idleCount: number
    canDraft: boolean
    tokenId: string | null
    lobbyRefresh: boolean
    selectedCard: string | null
    viewState: ViewState
    leagueLevel: string
    audioOn: boolean
    generatedCard: null | string
    /** Server sortOrders autoDraft for this wallet in this draft */
    autoDraft: boolean
}

export type mostRecentPlayerProps = {
    displayName: string
    ownerAddress: string
    pickNum: number
    playerId: string
    position: string
    round: number
    team: string
}